cs353-g3-backend-main/requirements.txt
django==3.2.16
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
pymongo==4.3.2
redis==4.3.4
psycopg2==2.9.4